backend/requirements.txt
Django
djangorestframework
django-cors-headers
channels
channels-redis
pillow
stripe
cryptography
django-user-agents
ua-parser
djangorestframework-simplejwt